✨ Create platform-specific SQLDelight setup
Using Koin and `expect class`, the code inside `commonMain` is
platform-agnostic. Now, every module implements its own platform driver.
For now, the prepopulation will only be executed in Android platform.
When the time comes, we refactor the code to also make it agnostic.
